Moved Archive and ArchivesDataSource to the global domain
