+++ VERA/VERAArchiving/VERAArchivingTests/DefaultArchiveRecordingsRepositoryTests.swift
@@ -8,7 +8,6 @@ import VERAArchiving
 import VERACore
 import VERADomain
 import VERATestHelpers
-import VERAVonage
 
 @Suite("Default archive recordings repository tests")
 struct DefaultArchiveRecordingsRepositoryTests {
