+++ VERA/VERAVonage/VERAVonage/HTTPArchivesDataSource.swift
@@ -3,7 +3,6 @@
 //
 
 import Foundation
-import VERAArchiving
 import VERADomain
 
 public final class HTTPArchivesDataSource: ArchivesDataSource {
@@ -34,5 +33,52 @@ public final class HTTPArchivesDataSource: ArchivesDataSource {
         let archivesResponse = try jsonDecoder.decode(RemoteArchivesResponse.self, from: response)
 
         return archivesResponse.archives.compactMap { $0.toDomain }
+    }
+}
+
+public struct RemoteArchivesResponse: Decodable {
+    public let archives: [RemoteArchive]
+    public let status: Int
+}
+
+public struct RemoteArchive: Decodable {
+    public let id: String
+    public let status: String
+    public let name: String
+    public let reason: String?
+    public let sessionId: String
+    public let applicationId: String
+    public let createdAt: TimeInterval
+    public let size: Int
+    public let duration: Int
+    public let outputMode: String
+    public let streamMode: String
+    public let hasAudio: Bool
+    public let hasVideo: Bool
+    public let hasTranscription: Bool
+    public let sha256sum: String
+    public let password: String
+    public let updatedAt: TimeInterval
+    public let multiArchiveTag: String
+    public let event: String
+    public let resolution: String
+    public let url: String?
+
+    public var toDomain: Archive? {
+        guard let uuid = UUID(uuidString: id) else {
+            return nil
+        }
+        return .init(
+            id: uuid,
+            name: name,
+            createdAt: Date(timeIntervalSince1970: createdAt),
+            status: ArchiveStatus(value: status),
+            url: url?.toURL)
+    }
+}
+
+extension String {
+    var toURL: URL? {
+        URL(string: self)
     }
 }
